some clippy fixes

+++ src/utils.rs
@@ -8,7 +8,7 @@ pub trait Round {
 impl Round for f64 {
     fn round_to(&self, decimal_places: u8) -> Self {
         let factor = 10.0f64.powi(decimal_places as i32);
-        (self * factor).round() as f64 / factor
+        (self * factor).round() / factor
     }
 }
